+++ UniVerServer.csproj
@@ -20,4 +20,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Models\CustomDataObjects\" />
+  </ItemGroup>
+
 </Project>
